Update to GPL-3.0 license.
Main difference is now that if anyone want to build something more with this,
then they need to open source their changes back to the project, which I thinks is fair.
This is so everyone can benefit from those improvements.
I you think this is wrong for some reason feel free to contact me,
I'm open to change it.