display: Optimize the way display power is handled.
This will turn off the display power completely after the watch has been left stationary for some time. As soon as the watch is moved, power will be applied and GC9A01 will be initialized, so that we can still have a fast wakeup when user wakes up the watch again.