Remove external_flash partition.
I don't get it why it's specified like this. When i read the docs it should indicate the unused
region of external flash. i.e. the region after all the static partition. By removing it, it's
automatically created and placed after the last pertions and spans until end of flash.