New display off tilt away feature to save power when user is not looking at the watch.
This is a new way to faster turn off the display when the user is not looking at the watch.
It uses the accelerometer to detect when the watch is tilted away from the user,
and turns off the display after a short delay if the watch remains in that position.